requirements/project.txt
# project.txt is a pip requirements file which describes the distributions
# required by your project to run.

--requirement=base.txt

--extra-index-url=http://dist.pinaxproject.com/dev/
# --extra-index-url=http://dist.pinaxproject.com/alpha/
# --extra-index-url=http://dist.pinaxproject.com/fresh-start/

--allow-all-external

# Put project-specific requirements here.
# See http://pip-installer.org/requirement-format.html for more information.

django-debug-toolbar==0.9.4
django-model-utils==2.2
-e git+git://github.com/pinax/django-mailer#egg=django-mailer

#  --------pinax dev
#django-openid==0.4.1
#  --------

#python-openid==2.2.5
django_compressor==1.3
pinax-blog==4.0.2
-e git+git://github.com/shvechikov/python-rtfng.git#egg=python-rtfng

django-selectable==0.7.0
django-forms-bootstrap==2.0.3.post1
django-uni-form==0.9.0

markedit==0.2.dev1

#django-social-auth==0.7.23
django-secure==1.0
raven==3.3.7
psycopg2==2.6
django-constance[database]==0.6
django-picklefield==0.3.1

django-redis-cache==0.9.7
redis==2.7.5
hiredis==0.1.1
#South==0.8.1

django-leaflet==0.14.1
django_osm_field==0.1.4
-e git+https://bitbucket.org/miurahr/django-restcms.git@e3b4600#egg=restcms

sphinx-intl

requests==2.6.0
Markdown==2.6.1

# hosting
gondor==1.2.5
gunicorn==19.3.0
opbeat==2.0.4
